// app/interfaces/InterfaceEjercicios.ts
export enum IMusculo{
    VOID = "",
    BICEP = "biceps",
    TRICEP = "triceps",
    ESPALDA = "espalda",
    DORSALES = "dorsales",
    CUADRICEPS = "cuadriceps",
    IZQUIOTIBIALES = "izquiotibiales",
    GEMELOS = "gemelos",
    HOMBRO = "hombro",
    ABDOMINALES = "abdominales",
    GLUTEO = "gluteo",
}

export enum IDia{
    VOID = "",
    LUNES = "lunes",
    MARTES = "martes",
    MIERCOLES = "miercoles",
    JUEVES = "jueves",
    VIERNES = "viernes",
    SABADO = "sabado",
    DOMINGO = "domingo",
}